chore: update extract evaluate pipeline
