+++ __tests__/evaluation/extract/evaluate.ts
@@ -4,15 +4,6 @@ import winston from 'winston';
 import { requestTboxLLM } from '../utils/common';
 import { GENRATE_RESULT_LOG_PATH, EVALUATE_ERROR_LOG_PATH, EVALUATE_RESULT_LOG_PATH } from './constants';
 
-const logger = winston.createLogger({
-  format: winston.format.json(),
-  defaultMeta: { service: 'evaluate-service' },
-  transports: [
-    new winston.transports.File({ filename: EVALUATE_ERROR_LOG_PATH, level: 'error' }),
-    new winston.transports.File({ filename: EVALUATE_RESULT_LOG_PATH, level: 'info' }),
-  ],
-});
-
 const prompt = `
   # 角色: 你是一个数据评测专家;
   # 任务: 我会给出一份日志数据，请你根据我的任务描述来评测，输出通过和不通过，以及给出通过或者不通过理由;
@@ -34,11 +25,11 @@ const prompt = `
   我给的日志信息如下:
 `;
 
-describe('evaluate cases', () => {
+export const evaluate = async () => {
   const content = fs.readFileSync(GENRATE_RESULT_LOG_PATH);
   const lines = content.toString().split('\n');
 
-  console.log(lines.length);
+  console.log('data length:', lines.length);
 
   if (fs.existsSync(EVALUATE_RESULT_LOG_PATH)) {
     fs.rmSync(EVALUATE_RESULT_LOG_PATH);
@@ -48,27 +39,52 @@ describe('evaluate cases', () => {
     fs.rmSync(EVALUATE_ERROR_LOG_PATH);
   }
 
-  it.each(lines)('test evaluate', async (line) => {
-    if (!line.length) return;
-    const data = JSON.parse(line);
-    const message = data.message;
+  const logger = winston.createLogger({
+    format: winston.format.json(),
+    defaultMeta: { service: 'evaluate-service' },
+    transports: [
+      new winston.transports.File({ filename: EVALUATE_ERROR_LOG_PATH, level: 'error' }),
+      new winston.transports.File({ filename: EVALUATE_RESULT_LOG_PATH, level: 'info' }),
+    ],
+  });
+
+  let success = 0;
+
+  for (const line of lines) {
+    if (!line.length) continue;
     try {
-      const res = await requestTboxLLM(prompt + '\n' + line);
-      const comment = JSON.parse(res);
-      logger.info({
-        pass: comment.pass,
-        reason: comment.reason,
-        input: message.input,
-        generate: message.generate,
-        source: message.source,
-      });
+      const data = JSON.parse(line);
+      const { message } = data;
+      try {
+        const res = await requestTboxLLM(`${prompt}\n${line}`);
+        const comment = JSON.parse(res);
+        logger.info({
+          pass: comment.pass,
+          reason: comment.reason,
+          input: message.input,
+          generate: message.generate,
+          source: message.source,
+        });
+        success++;
+      } catch (e) {
+        logger.error({
+          error: e.message,
+          input: message.input,
+          generate: message.generate,
+          source: message.source,
+        });
+      }
     } catch (e) {
       logger.error({
-        error: e.message,
-        input: message.input,
-        generate: message.generate,
-        source: message.source,
+        error: 'parse error',
+        data: line,
       });
+      continue;
     }
-  });
-});
+  }
+
+  return {
+    success,
+    total: lines.length,
+  };
+};
