+++ jest.setup.js
@@ -1,5 +1,4 @@
 require('dotenv').config({
   path: ['.env.local', '.env']
 });
-
 jest.setTimeout(30000);
