+++ package.json
@@ -30,7 +30,7 @@
     "test": "jest __tests__/unit",
     "evaluate": "run-s evaluate:*",
     "evaluate:advise": "jest __tests__/evaluation/advise --verbose --runInBand",
-    "evaluate:extract": "jest __tests__/evaluation/extract --verbose --runInBand",
+    "evaluate:extract": "jest __tests__/evaluation/extract --verbose",
     "lint-staged": "lint-staged"
   },
   "lint-staged": {
